refactor bigdecimal

+++ src/main/java/com/example/fairshareapp/model/dto/CrearGastoDTO.java
@@ -1,6 +1,6 @@
 package com.example.fairshareapp.model.dto;
 
-import com.example.fairshareapp.model.entity.ReglaDivision;
+import com.example.fairshareapp.model.enums.ReglaDivision;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
@@ -13,6 +13,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -32,7 +33,7 @@ public class CrearGastoDTO {
 
     @NotNull(message = "El monto es obligatorio")
     @Positive(message = "El monto del gasto debe ser un valor positivo")
-    private Double monto;
+    private BigDecimal monto;
 
     private LocalDate fecha;
